Don't remember what I was working on here, but it looks like it was just fixing the script to upgrade to 3.2 version of the SDK.  The script is so far from stable that I don't think it matters.
I want to check this in so that I can play with creating a Gentoo-based system.  I like Gentoo because I don't have to keep the C API aligned with every application.  Gentoo always compiles applications upon installation.  This requires a lot of CPU/drive speed, but I'm hopeful I can work towards parallelizing this using EC2/S3.  It'll be nice to be able to move to bleeding edge stuff and know I have source installed for most components.